requirements.txt
nose
cocos2d
pyglet